Moving Channel and Message into their own class
