+++ lib/java/os-environment/src/main/java/org/enso/os/environment/jni/JVM.java
@@ -3,18 +3,13 @@ package org.enso.os.environment.jni;
 import java.io.File;
 import java.io.IOException;
 import java.lang.foreign.Arena;
-import java.lang.foreign.FunctionDescriptor;
-import java.lang.foreign.Linker;
 import java.lang.foreign.MemorySegment;
-import java.lang.foreign.ValueLayout;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Function;
 import org.enso.common.Platform;
 import org.enso.persist.Persistance;
-import org.graalvm.nativeimage.CurrentIsolate;
-import org.graalvm.nativeimage.ImageInfo;
 import org.graalvm.nativeimage.IsolateThread;
 import org.graalvm.nativeimage.StackValue;
 import org.graalvm.nativeimage.UnmanagedMemory;
@@ -71,7 +66,7 @@ public final class JVM {
   }
 
   static <R> R executeImpl(
-      Persistance.Pool pool, Message<R> msg, Function<MemorySegment, Long> send) {
+      Persistance.Pool pool, Channel.Message<R> msg, Function<MemorySegment, Long> send) {
     try (var arena = Arena.ofConfined()) {
       var bytes = pool.write(msg, null);
       var memory = arena.allocate(Math.max(bytes.length, 4096));
@@ -104,127 +99,6 @@ public final class JVM {
           CCharPointer.class,
           long.class);
 
-  /** Channel connects two JVMs. */
-  public static final class Channel {
-    /** persistance pool associated with this channel object */
-    private final Persistance.Pool pool;
-
-    private final JNI.JNIEnv env;
-    private final long isolate;
-    private final long callbackFn;
-
-    /* private */ Channel(Persistance.Pool pool, JNI.JNIEnv env) {
-      this.pool = pool;
-      this.env = env;
-      this.isolate = -1;
-      this.callbackFn = -1;
-    }
-
-    /* private */ Channel(Persistance.Pool pool, long isolate, long callbackFn) {
-      if (ImageInfo.inImageCode()) {
-        throw new IllegalStateException("Only usable in HotSpot");
-      }
-      this.pool = pool;
-      this.env = null;
-      this.isolate = isolate;
-      this.callbackFn = callbackFn;
-    }
-
-    /**
-     * <em>Executes a message</em> in the other JVM. The message is any subclass of {@link Message}
-     * registered for persistance via {@link Persistable @Persistable} annotation into the {@link
-     * Persistance.Pool pool associated with this JVM}. The result (which is of type {@code R}) also
-     * has to be registered for serde.
-     *
-     * @param msg the message that gets serialized, transfered into the other JVM, deserialized on
-     *     the other side and {@link Message#evaluate() evaluated} there
-     * @param <R> the type of result we expect the message to return
-     * @return the value gets computed via {@link Message#evaluate()} in the other JVM and then it
-     *     gets serialized and transfered back to us. Deserialized and the value is then returned
-     *     from this method
-     */
-    public final <R> R execute(Message<R> msg) {
-      if (this.isolate == -1) {
-        return executeImpl(pool, msg, (memory) -> toHotSpotMessage(env, memory));
-      } else {
-        var fnCallbackAddress = MemorySegment.ofAddress(callbackFn);
-        var fnDescriptor =
-            FunctionDescriptor.of(
-                ValueLayout.JAVA_LONG,
-                ValueLayout.ADDRESS,
-                ValueLayout.ADDRESS,
-                ValueLayout.JAVA_LONG);
-        var fnHandle = Linker.nativeLinker().downcallHandle(fnCallbackAddress, fnDescriptor);
-        return executeImpl(
-            pool,
-            msg,
-            (seg) -> {
-              Object res = -1L;
-              try {
-                var isoRef = MemorySegment.ofAddress(isolate);
-                res = fnHandle.invoke(isoRef, seg, seg.byteSize());
-              } catch (Throwable ex) {
-                ex.printStackTrace();
-              }
-              return (long) res;
-            });
-      }
-    }
-
-    private static long toHotSpotMessage(JNI.JNIEnv e, MemorySegment segment) {
-      var classNameWithSlashes = "org/enso/os/environment/jni/JVMPeer";
-      var methodName = "handle";
-      try (var classInC = CTypeConversion.toCString(classNameWithSlashes);
-          var methodInC = CTypeConversion.toCString(methodName);
-          var signatureInC = CTypeConversion.toCString("(JJJJ)J"); ) {
-        var fn = e.getFunctions();
-        var clazz = fn.getFindClass().call(e, classInC.get());
-        assert clazz.isNonNull() : "Class not found " + classNameWithSlashes;
-        var method = fn.getGetStaticMethodID().call(e, clazz, methodInC.get(), signatureInC.get());
-        assert method.isNonNull() : "method not found in " + classNameWithSlashes;
-        var address = segment.address();
-        assert address > 0 : "We need an address";
-        var arg = StackValue.get(4, JNI.JValue.class);
-        arg.addressOf(0).setLong(CurrentIsolate.getCurrentThread().rawValue());
-        arg.addressOf(1).setLong(CALLBACK_FN.getFunctionPointer().rawValue());
-        arg.addressOf(2).setLong(address);
-        arg.addressOf(3).setLong(segment.byteSize());
-        var replySize = fn.getCallStaticLongMethodA().call(e, clazz, method, arg);
-        return replySize;
-      }
-    }
-  }
-
-  /**
-   * Subclasses of message denote a computational task to be performed in the "other JVM".
-   *
-   * @param <R> type of the return value
-   */
-  public abstract static class Message<R> {
-    private final Class<R> replyType;
-
-    /**
-     * Constructor for subclasses. Use it as {@code super(Integer.class)} to specify the reply type
-     * of the exception which is then returned from the {@link
-     * #execute(org.enso.os.environment.jni.JVM.Message)} method.
-     *
-     * @param replyType the type of the reply
-     */
-    protected Message(Class<R> replyType) {
-      this.replyType = replyType;
-    }
-
-    /**
-     * Evaluates the exception. Invoked in the other JVM.
-     *
-     * @param channel allows sending messages to the other JVM
-     * @return the result of the evaluation or {@code null}
-     * @throws Throwable the computation may yield exceptions or errors which are then transferred
-     *     back to the callee JVM
-     */
-    protected abstract R evaluate(Channel channel) throws Throwable;
-  }
-
   /**
    * Executes main method of provided class
    *
@@ -233,7 +107,7 @@ public final class JVM {
    */
   public final void executeMain(String classNameWithSlashes, String... args) {
     var msg = new JVMPeer.ExecuteMainClass(classNameWithSlashes, List.of(args));
-    var channel = new Channel(JVMPeer.POOL, env());
+    org.enso.os.environment.jni.Channel channel = new Channel(JVMPeer.POOL, env());
     channel.execute(msg);
   }
