+++ lib/java/os-environment/src/main/java/org/enso/os/environment/jni/JVMPeer.java
@@ -6,7 +6,6 @@ import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import org.enso.os.environment.jni.JVM.Channel;
 import org.enso.persist.Persistable;
 import org.enso.persist.Persistance;
 
@@ -18,7 +17,7 @@ final class JVMPeer {
   /** called via JNI from SVM to HotSpot */
   static long handle(long threadId, long callbackFn, long address, long size) {
     try {
-      var channel = new JVM.Channel(JVMPeer.POOL, threadId, callbackFn);
+      var channel = new Channel(JVMPeer.POOL, threadId, callbackFn);
       return handleWithChannel(channel, address, size);
     } catch (Throwable t) {
       // TBD: proper handling of exceptions is needed
@@ -31,7 +30,7 @@ final class JVMPeer {
     var seg = MemorySegment.ofAddress(address).reinterpret(size);
     var buf = seg.asByteBuffer();
     var ref = POOL.read(buf, null);
-    var msg = ref.get(JVM.Message.class);
+    var msg = ref.get(Channel.Message.class);
     var res = msg.evaluate(channel);
     var bytes = Persistables.POOL.write(res, null);
     seg.copyFrom(MemorySegment.ofArray(bytes));
@@ -39,7 +38,7 @@ final class JVMPeer {
   }
 
   @Persistable(id = 432001)
-  public static final class ExecuteMainClass extends JVM.Message<Void> {
+  public static final class ExecuteMainClass extends Channel.Message<Void> {
     private final String mainClassWithSlashes;
     private final List<String> args;
