+++ lib/java/os-environment/src/test/java/org/enso/os/environment/jni/LoadClassTest.java
@@ -138,7 +138,7 @@ public class LoadClassTest {
   public void computeFactorialViaMessages() throws Exception {
     TestMain.CORRECT_RESULTS.clear();
     assertEquals("Results are empty", 0, TestMain.CORRECT_RESULTS.size());
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var gen = new Random();
     var n = 0L;
     for (var i = 0; i < 5; i++) {
@@ -155,7 +155,7 @@ public class LoadClassTest {
 
   @Test
   public void computeFactorialViaSingleMessage() throws Exception {
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var gen = new Random();
     var n = 0L;
     for (var i = 0; i < 5; i++) {
@@ -168,14 +168,14 @@ public class LoadClassTest {
 
   @Test
   public void backAndForthFactorialOne() throws Exception {
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var fac = channel.execute(new TestMain.CountDownAndReturn(1, 1));
     assertEquals(1, fac.longValue());
   }
 
   @Test
   public void backAndForthFactorialTwo() throws Exception {
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var fac = channel.execute(new TestMain.CountDownAndReturn(2, 1));
     assertEquals(2, fac.longValue());
   }
@@ -183,7 +183,7 @@ public class LoadClassTest {
   @Ignore
   @Test
   public void backAndForthFactorialThree() throws Exception {
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var fac = channel.execute(new TestMain.CountDownAndReturn(3, 1));
     assertEquals(6, fac.longValue());
   }
@@ -191,7 +191,7 @@ public class LoadClassTest {
   @Ignore
   @Test
   public void backAndForthFactorialFour() throws Exception {
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var fac = channel.execute(new TestMain.CountDownAndReturn(4, 1));
     assertEquals(24, fac.longValue());
   }
@@ -199,7 +199,7 @@ public class LoadClassTest {
   @Ignore
   @Test
   public void backAndForthFactorialFive() throws Exception {
-    var channel = new JVM.Channel(JVMPeer.POOL, env());
+    var channel = new Channel(JVMPeer.POOL, env());
     var fac = channel.execute(new TestMain.CountDownAndReturn(5, 1));
     assertEquals(120, fac.longValue());
   }
